feature: restructure the experiences section
