+++ src/components/SocialMedia/style.css
@@ -27,3 +27,9 @@
     height: 100px;
     background: var(--primary-transparency);
 }
+
+@media screen and (max-width: 768px) {
+    .fixed {
+        display: none;
+    }
+}
