+++ src/data/experiences.jsx
@@ -14,7 +14,8 @@ export const experiences_pt = [
     title: "Desenvolvedora Front-end",
     level: "Pleno",
     company: "Outliant",
-    location: "California, US (remoto)",
+    location: "California, US",
+    workLocationType: "remoto",
     period: "12/2024 — presente",
     duration: "1 ano",
     type: "work",
@@ -41,7 +42,8 @@ export const experiences_pt = [
     title: "Desenvolvedora Front-end",
     level: "Pleno",
     company: "FTD Educação",
-    location: "São Paulo, BR (remoto)",
+    location: "São Paulo, BR",
+    workLocationType: "remoto",
     period: "10/2023 — 06/2025",
     duration: "2 anos",
     type: "work",
@@ -69,7 +71,8 @@ export const experiences_pt = [
     title: "Desenvolvedora Front-end",
     level: "Júnior",
     company: "Agência Choveu",
-    location: "São Paulo, BR (remoto)",
+    location: "São Paulo, BR",
+    workLocationType: "remoto",
     period: "03/2022 — 03/2024",
     duration: "2 anos",
     type: "work",
@@ -100,7 +103,8 @@ export const experiences_pt = [
     period: "08/2022 — 08/2024",
     duration: "2 anos",
     school: "FIAP",
-    location: "São Paulo, BR (presencial)",
+    location: "São Paulo, BR",
+    workLocationType: "presencial",
     link: "https://www.fiap.com.br/",
     summary:
       "Durante minha graduação na FIAP, aprimorei minhas habilidades em desenvolvimento web, mobile, banco de dados e inteligência artificial. Além disso, tenho a chance de criar projetos práticos em parceria com empresas incríveis, como a B3 e a Level Group.",
@@ -128,7 +132,8 @@ export const experiences_pt = [
     period: "02/2019 — 11/2021",
     duration: "3 anos",
     school: "ETEC de Guaianazes",
-    location: "São Paulo, BR (presencial)",
+    location: "São Paulo, BR",
+    workLocationType: "presencial",
     link: "https://www.etecguaianazes.com.br/",
     summary:
       "Na ETEC eu descobri a programação como profissão. Durante os 3 anos que estive lá construí uma base sólida em lógica de programação e desenvolvimento. Também tive experiências incríveis como as maratonas de programação e visitas técnicas em empresas de grande porte.",
@@ -152,7 +157,8 @@ export const experiences_en = [
     ...experiences_pt[0],
     id: uuid(),
     title: "Frontend Software Engineer",
-    location: "California, US (remote)",
+    location: "California, US",
+    workLocationType: "remote",
     level: "Mid-level",
     period: "12/2024 — present",
     duration: "1 year",
@@ -169,7 +175,8 @@ export const experiences_en = [
     ...experiences_pt[1],
     id: uuid(),
     title: "Frontend Developer",
-    location: "São Paulo, BR (remote)",
+    location: "São Paulo, BR",
+    workLocationType: "remote",
     level: "Mid-level",
     duration: "2 years",
     summary:
@@ -186,7 +193,8 @@ export const experiences_en = [
     ...experiences_pt[2],
     id: uuid(),
     title: "Frontend Developer",
-    location: "São Paulo, BR (remote)",
+    location: "São Paulo, BR",
+    workLocationType: "remote",
     level: "Junior",
     duration: "2 years",
     summary:
@@ -202,7 +210,8 @@ export const experiences_en = [
     ...experiences_pt[3],
     id: uuid(),
     title: "Systems Analysis and Development",
-    location: "São Paulo, BR (on-site)",
+    location: "São Paulo, BR",
+    workLocationType: "on-site",
     level: "Bachelor's Degree",
     duration: "2 years",
     summary:
@@ -217,7 +226,8 @@ export const experiences_en = [
     ...experiences_pt[4],
     id: uuid(),
     title: "Systems Development",
-    location: "São Paulo, BR (on-site)",
+    location: "São Paulo, BR",
+    workLocationType: "on-site",
     level: "Technician",
     duration: "3 years",
     summary:
