+++ src/index.css
@@ -231,7 +231,7 @@ a {
 
 @media screen and (max-width: 768px) {
   .container{
-    padding: 0 3.5em;
+    padding: 0 2rem;
   }
 
   .column, #hero .text{
